AMS-5 antiduplication changes via concern and templates
