+++ app/controllers/doctor_admin/application_controller.rb
@@ -1,7 +1,7 @@
 class DoctorAdmin::ApplicationController < ApplicationController
   before_action :authenticate_doctor_admin_doctor_user!
 
-  layout 'doctor_admin'
+  layout 'dashboard'
 
   helper_method :current_user, :namespace
