+++ app/controllers/doctor_admin/passwords_controller.rb
@@ -1,3 +1,3 @@
 class DoctorAdmin::PasswordsController < Devise::PasswordsController
-  layout 'doctor_auth'
+  layout 'auth'
 end
