+++ app/controllers/doctor_admin/registrations_controller.rb
@@ -1,5 +1,5 @@
 class DoctorAdmin::RegistrationsController < Devise::RegistrationsController
-  layout 'doctor_auth'
+  layout 'auth'
 
   private
