+++ app/controllers/patient_admin/passwords_controller.rb
@@ -1,3 +1,3 @@
 class PatientAdmin::PasswordsController < Devise::PasswordsController
-  layout 'patient_auth'
+  layout 'auth'
 end
