+++ app/controllers/patient_admin/registrations_controller.rb
@@ -1,5 +1,5 @@
 class PatientAdmin::RegistrationsController < Devise::RegistrationsController
-  layout 'patient_auth'
+  layout 'auth'
 
   private
