+++ app/controllers/patient_admin/sessions_controller.rb
@@ -1,5 +1,5 @@
 class PatientAdmin::SessionsController < Devise::SessionsController  
-  layout 'patient_auth'
+  layout 'auth'
 
   def new
     super{ |resource| return render 'shared/sessions/new' }
