+++ app/views/doctor_admin/after_signup/personal.html.slim
@@ -1,42 +1,1 @@
-h1.text-center Personal
-h1 | Please fill this states fo rmore information 
-section.pos-r
-  = simple_form_for @personal, url: wizard_path, method: :put  do |f|
-    fieldset
-      .row
-        .col-lg-6.col-md-6.col-sm-12.col-xs-12
-          .form-group.no-side-margins
-            label= t 'registration.common.fields.address'
-            = f.input_field :address, required: true, autofocus: true, tabindex: 1
-        .col-lg-6.col-md-6.col-sm-12.col-xs-12
-          .form-group.no-side-margins
-            label= t 'registration.common.fields.city'
-            = f.input_field :city, required: true, tabindex: 2
-      .row
-        .col-lg-6.col-md-6.col-sm-12.col-xs-12
-          .form-group.no-side-margins
-            label= t 'registration.common.fields.state'
-            = f.input_field :clinic, required: true, tabindex: 3
-      .row
-        .col-lg-6.col-md-6.col-sm-12.col-xs-12
-          .form-group.no-side-margins
-            label= t 'registration.common.fields.phone'
-            = f.input_field :phone, class: 'js-phone-input', required: true, tabindex: 4
-        .col-lg-6.col-md-6.col-sm-12.col-xs-12
-          .form-group.no-side-margins
-            label= t 'registration.common.fields.phone'
-            = f.input_field :speciality, class: 'js-phone-input', required: true, tabindex: 5
-      .row 
-        .col-lg-6.col-md-6.col-sm-12.col-xs-12
-          .form-group.no-side-margins
-            label= t 'registration.common.fields.phone'
-            = f.input_field :license_id_photo, class: 'js-phone-input', required: true, tabindex: 6
-        .col-lg-6.col-md-6.col-sm-12.col-xs-12
-          .form-group.no-side-margins
-            label= t 'registration.common.fields.phone'
-            = f.input_field :face_photo, class: 'js-phone-input', required: true, tabindex: 7
-      .row
-        .col-sm-12
-          .form-group.no-side-margins
-            = f.submit 'Sign Up', class: 'btn btn-success'
-            
+= render 'shared/after_signup/personal'
