+++ app/views/layouts/auth.html.slim
