--- app/views/layouts/patient_auth.html.slim
@@ -1,28 +1,0 @@
-doctype html
-html lang="en"
-  head
-    meta charset="utf-8"
-    meta http-equiv="X-UA-Compatible" content="IE=Edge,chrome=1"
-    meta name="viewport" content="width=device-width, initial-scale=1.0"
-    title= content_for?(:title) ? yield(:title) : "Heally"
-    = csrf_meta_tags
-
-      /! Le HTML5 shim, for IE6-8 support of HTML elements
-    /[if lt IE 9]
-      = javascript_include_tag "//cdnjs.cloudflare.com/ajax/libs/html5shiv/3.6.1/html5shiv.js"
-    = stylesheet_link_tag 'application', media: 'all'
-    = favicon_link_tag 'apple-touch-icon-144x144.png', rel: 'apple-touch-icon-precomposed', type: 'image/png', sizes: '144x144'
-    = favicon_link_tag 'apple-touch-icon-114x114.png', rel: 'apple-touch-icon-precomposed', type: 'image/png', sizes: '114x114'
-    = favicon_link_tag 'apple-touch-icon-72x72.png', rel: 'apple-touch-icon-precomposed', type: 'image/png', sizes: '72x72'
-    = favicon_link_tag 'apple-touch-icon-precomposed.png', rel: 'apple-touch-icon-precomposed', type: 'image/png'
-    = favicon_link_tag 'favicon.ico', rel: 'shortcut icon'
-    = javascript_include_tag 'application'
-      
-  body.auth.gray-bg
-    = yield(:modals)
-    .wrapper
-      .page-wrapper
-        = render 'shared/auth_header'
-        .container.page-content
-          = yield
-        = render 'shared/auth_footer'
